backend/requirements.txt
fastapi==0.115.2
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
requests==2.32.0
python-multipart==0.0.6
